引数をuint32からuint16へ変更

+++ src/PE3VM3/App/Central/esw/EthSwt/Config/EthSwt_SWIC_Cfg.c
@@ -199,14 +199,14 @@ Std_ReturnType EthSwt_SWIC_Cfg_AllowRelay(void)
 /* -------------------------------------------------------------------------- */
 /* Config for EthSwt_SWIC_Reg.c                                               */
 /* -------------------------------------------------------------------------- */
-void EthSwt_SWIC_Cfg_WaitUS(const uint32 waitUSTime)			                                /* waitUSの値域はD_ETHSWT_SWIC_CLOCK_PER_USによる。今回は0~10,737,418 */
+void EthSwt_SWIC_Cfg_WaitUS(const uint16 waitUSTime)
 {
 	const uint32	startTime = LIB_GetFreeRunCount1us();
 	uint32			i;
 
-	for (i = 0U; i < (waitUSTime * D_ETHSWT_SWIC_CLOCK_PER_US); i++) {                          /* 1ループ1クロック前提（実際はそれ以上のクロックがある）にガード実装 */
+	for (i = 0U; i < ((uint32)waitUSTime * D_ETHSWT_SWIC_CLOCK_PER_US); i++) {      /* 1ループ1クロック前提（実際はそれ以上のクロックがある）にガード実装 */
 		const uint32	nowTime = LIB_GetFreeRunCount1us();
-		if ((nowTime - startTime) > waitUSTime)	{ break; }                                      /* ラップアラウンドはC規格上問題なし */
+		if ((nowTime - startTime) > (uint32)waitUSTime)	{ break; }                  /* ラップアラウンドはC規格上問題なし */
 	}
 
 	return;
